Check for duplicate stdin usage on read instead of arg parsing (#10)
This change relaxes the check for duplicate usage of `stdin` on arg
declarations (error at runtime if any two args use `MaybeStdin` or
`FileOrStdin`) and instead only check for duplicated `stdin` usage
when these args are accessed.

This allows usages of args that may be mutually exclusive (E.g. under
different subcommands) or use the `global=true` clap option (as reported
in #9)

If a tool happens to accept multiple args that can be Stdin, the CLI user
will only see an error if they actually try to use `stdin` for values
twice.